Remove func-tools-common.
