--- Tasks/Common/func-tools-common/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "func-tools-common",
-  "version": "1.0.0",
-  "description": "Common Library for Azure func tools",
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/Microsoft/azure-pipelines-tasks.git"
-  },
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Microsoft/azure-pipelines-tasks/issues"
-  },
-  "homepage": "https://github.com/Microsoft/azure-pipelines-tasks#readme",
-  "dependencies": {
-    "azure-pipelines-task-lib": "2.8.0",
-    "azure-pipelines-tool-lib": "0.12.0"
-  }
-}
