--- Tasks/Common/func-tools-common/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es6",
-        "declaration": true,
-        "noImplicitAny": false,
-        "sourceMap": false
-    }
-}
