--- Tasks/Common/func-tools-common/typings.json
@@ -1,5 +1,0 @@
-{
-    "globalDependencies": {
-        "node": "registry:dt/node#6.0.0+20160920093002"
-    }
-}
