--- Tasks/FuncToolsInstallerV0/make.json
@@ -1,16 +1,0 @@
-{
-    "common": [{
-        "module": "../Common/func-tools-common",
-        "type": "node",
-        "dest" : "./",
-        "compile" : true
-    }],
-    "rm": [
-        {
-            "items": [
-                "node_modules/func-tools-common/node_modules/azure-pipelines-task-lib"
-            ],
-            "options": "-Rf"
-        }
-    ]
-}
