+++ Tasks/FuncToolsInstallerV0/task.json
@@ -14,7 +14,7 @@
     "version": {
         "Major": 0,
         "Minor": 163,
-        "Patch": 2
+        "Patch": 3
     },
     "preview": true,
     "demands": [],
